mark task as complete and implement
