+++ src/index.js
@@ -1,5 +1,6 @@
 // import _ from 'lodash';
 import './style.css';
+import './bootstrap.min.css';
 import { add, clear, windowLoad } from './task.js';
 
 add();
