+++ src/task.js
@@ -1,6 +1,8 @@
+/* eslint-disable no-undef */
 /* eslint-disable import/no-mutable-exports */
 // eslint-disable-next-line import/no-cycle
 import addDataTolist from './addDataTolist.js';
+import completedTask from './completedTask.js';
 
 export const data = JSON.parse(localStorage.getItem('data')) || [];
 const addItem = document.getElementById('add-item');
@@ -12,33 +14,62 @@ export function displayList(data) {
     data.forEach((d) => {
       const div = document.createElement('div');
       div.classList.add('task-items');
+      div.setAttribute('id', `cont-${d.index}`);
 
       let content = '';
-      content += `<div class="task-item list" id="${d.index + 1}"><div class="box" id="${d.index}">
+      content += `<div class="task-item list" id="div-${d.index}"><div class="box" id="${d.index}">
       <input type="checkbox" id="task-item" name="task" value="task"  ${d.completed ? 'checked' : ''} class="checkBox">
       <div class="description ${d.completed ? 'checked' : ''}" contenteditable="${!d.completed}"> ${d.description}</div>
-      </div><i class="fas fa-trash delete"></i></div>
+      </div>
+      <i class="fas fa-trash delete d-none" id="del-${d.index}"></i>
+      <i class="fas fa-ellipsis-v ellipsis" id="el-${d.index}"></i>
+      </div>
       <div class="grey-rule"></div>`;
       div.innerHTML = content;
       listItems.appendChild(div);
     });
   }
 
-  const completedTask = (e) => {
-    const data = localStorage.getItem('data');
-    const dataArray = JSON.parse(data);
-    const listObject = dataArray.find((x) => x.index === Number(e.target.parentElement.id));
-    const index = dataArray.indexOf(listObject);
-    if (!dataArray[index].completed) {
-      dataArray[index].completed = 1;
-      e.target.parentElement.children[1].classList.add('checked');
-      localStorage.setItem('data', JSON.stringify(dataArray));
-    } else {
-      dataArray[index].completed = 0;
-      e.target.parentElement.children[1].classList.remove('checked');
-      localStorage.setItem('data', JSON.stringify(dataArray));
-    }
-  };
+  data.forEach((d) => {
+    const taskItems = document.getElementById(`cont-${d.index}`);
+    const taskItem = document.getElementById(`div-${d.index}`);
+    const trashIcon = document.getElementById(`del-${d.index}`);
+    const ellipsis = document.getElementById(`el-${d.index}`);
+
+    taskItem.addEventListener('focusin', () => {
+      ellipsis.classList.add('d-none');
+      trashIcon.classList.remove('d-none');
+      trashIcon.classList.add('d-block');
+      // taskItems.classList.toggle('itemSelected');
+    });
+
+    taskItem.addEventListener('focusout', () => {
+      setTimeout(() => {
+        ellipsis.classList.add('d-block');
+        ellipsis.classList.remove('d-none');
+        trashIcon.classList.add('d-none');
+        // taskItems.classList.toggle('listSelected');
+      }, 100);
+    });
+  });
+
+  //   console.log(taskItems[i]);
+
+  // const completedTask = (e) => {
+  //   const data = localStorage.getItem('data');
+  //   const dataArray = JSON.parse(data);
+  //   const listObject = dataArray.find((x) => x.index === Number(e.target.parentElement.id));
+  //   const index = dataArray.indexOf(listObject);
+  //   if (!dataArray[index].completed) {
+  //     dataArray[index].completed = 1;
+  //     e.target.parentElement.children[1].classList.add('checked');
+  //     localStorage.setItem('data', JSON.stringify(dataArray));
+  //   } else {
+  //     dataArray[index].completed = 0;
+  //     e.target.parentElement.children[1].classList.remove('checked');
+  //     localStorage.setItem('data', JSON.stringify(dataArray));
+  //   }
+  // };
   const checkBox = document.querySelectorAll('.checkBox');
   checkBox.forEach((btn) => {
     btn.addEventListener('change', (e) => {
